@@ -3,19 +3,18 @@
 	"description": "Web interface for Intuition.",
 	"license": "MIT",
 	"require": {
-		"php": ">=5.5",
+		"php": ">=7.0",
 		"krinkle/toollabs-base": "^0.8.1",
-		"krinkle/intuition": "^0.6.0"
+		"krinkle/intuition": "^1.1.0"
 	},
 	"require-dev": {
-		"jakub-onderka/php-parallel-lint": "0.*",
-		"squizlabs/php_codesniffer": "2.3.*",
-		"mediawiki/mediawiki-codesniffer": "0.5.0"
+		"jakub-onderka/php-parallel-lint": "1.0.0",
+		"mediawiki/mediawiki-codesniffer": "21.0.0"
 	},
 	"scripts": {
 		"test": [
 			"parallel-lint . --exclude libs --exclude vendor --exclude node_modules",
-			"phpcs . -s --extensions=php --standard=tests/codesniffer --ignore=vendor/*,node_modules/*,libs/*",
+			"phpcs -p -s",
 			"npm install && npm test"
 		]
 	}
